# api/openapi.yaml
openapi: 3.0.3
info:
  title: Insdr Messenger API
  description: Automatic message sending system with scheduling capabilities
  version: 1.0.0
  contact:
    name: API Support
    email: support@insdr.com

servers:
  - url: http://localhost:8080
    description: Local development server

paths:
  /scheduler/start:
    post:
      tags:
        - Scheduler
      summary: Start automatic message sending
      description: Starts the automatic message sending scheduler
      operationId: startScheduler
      responses:
        '200':
          description: Scheduler started successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SchedulerResponse'
        '409':
          description: Scheduler is already running
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /scheduler/stop:
    post:
      tags:
        - Scheduler
      summary: Stop automatic message sending
      description: Stops the automatic message sending scheduler
      operationId: stopScheduler
      responses:
        '200':
          description: Scheduler stopped successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SchedulerResponse'
        '409':
          description: Scheduler is not running
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /messages/sent:
    get:
      tags:
        - Messages
      summary: Get list of sent messages
      description: Retrieves a list of all sent messages with pagination support
      operationId: getSentMessages
      parameters:
        - name: page
          in: query
          description: Page number for pagination
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          description: Number of items per page
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: List of sent messages retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageListResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /health:
    get:
      tags:
        - Health
      summary: Health check endpoint
      description: Check if the service is healthy and running
      operationId: healthCheck
      responses:
        '200':
          description: Service is healthy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
        '503':
          description: Service is unhealthy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'

components:
  schemas:
    SchedulerResponse:
      type: object
      required:
        - status
        - message
      properties:
        status:
          type: string
          enum: [started, stopped]
          description: Current status of the scheduler
        message:
          type: string
          description: Status message
          example: "Scheduler started successfully"

    MessageListResponse:
      type: object
      required:
        - messages
        - pagination
      properties:
        messages:
          type: array
          items:
            $ref: '#/components/schemas/Message'
        pagination:
          $ref: '#/components/schemas/Pagination'

    Message:
      type: object
      required:
        - id
        - phone_number
        - content        - sent_at
        - status
      properties:
        id:
          type: integer
          format: int64
          description: Unique message identifier
        phone_number:
          type: string
          description: Recipient phone number
          example: "+905551111111"
        content:
          type: string
          description: Message content
          maxLength: 160
          example: "Insdr - Project"
        sent_at:
          type: string
          format: date-time
          description: Timestamp when the message was sent
        status:
          type: string
          enum: [pending, sent, failed]
          description: Message sending status
        message_id:
          type: string
          description: External message ID from webhook response
          example: "67f2f8a8-ea58-4ed0-a6f9-ff217df4d849"
          nullable: true
        error:
          type: string
          description: Error message if sending failed
          nullable: true

    Pagination:
      type: object
      required:
        - current_page
        - total_pages
        - total_items
        - items_per_page
      properties:
        current_page:
          type: integer
          description: Current page number
          minimum: 1
        total_pages:
          type: integer
          description: Total number of pages
          minimum: 0
        total_items:
          type: integer
          description: Total number of items
          minimum: 0
        items_per_page:
          type: integer
          description: Number of items per page
          minimum: 1
          maximum: 100

    HealthResponse:
      type: object
      required:
        - status
        - timestamp
      properties:
        status:
          type: string
          enum: [healthy, unhealthy, degraded]
          description: Service health status
        timestamp:
          type: string
          format: date-time
          description: Current server timestamp
        scheduler_status:
          type: string
          enum: [running, stopped]
          description: Current scheduler status
          nullable: true
        database_status:
          type: string
          enum: [connected, disconnected]
          description: Database connection status
          nullable: true
        redis_status:
          type: string
          enum: [connected, disconnected]
          description: Redis connection status
          nullable: true
        circuit_breaker_status:
          type: string
          description: Circuit breaker statistics
          nullable: true
        circuit_breaker_state:
          type: string
          enum: [closed, open, half-open]
          description: Current circuit breaker state
          nullable: true

    ErrorResponse:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: Error code
          example: "SCHEDULER_ALREADY_RUNNING"
        message:
          type: string
          description: Human-readable error message
          example: "Scheduler is already running"
        timestamp:
          type: string
          format: date-time
          description: Timestamp of the error
          nullable: true

tags:
  - name: Scheduler
    description: Operations for controlling the message scheduler
  - name: Messages
    description: Operations for managing messages
  - name: Health
    description: Health check operations